Fixed session and ByteCoin calculation

+++ database/migrations/2019_09_20_123039_add_byte_coins_column_to_computers_table.php
@@ -14,7 +14,7 @@ class AddByteCoinsColumnToComputersTable extends Migration
     public function up()
     {
         Schema::table('computers', function (Blueprint $table) {
-            $table->unsignedBigInteger('byte_coins')->default(0);
+            $table->float('byte_coins')->default(0);
         });
     }
